requirements.txt
google-adk
google-generativeai
pypdf
python-dotenv
reportlab
